backend/core/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
httpx
pytest
# supertokens-python
